226.Typing-Button-Component
button reformed to ts file, ButtonHTMLAttributes used for extending button attributes
