+++ crwn-clothing/package.json
@@ -14,6 +14,7 @@
     "@types/node": "^22.13.10",
     "@types/react": "^19.0.10",
     "@types/react-dom": "^19.0.4",
+    "@types/styled-components": "^5.1.34",
     "dotenv": "^16.4.7",
     "firebase": "^11.4.0",
     "react": "^19.0.0",
